requirements.txt
ultralytics==8.3.188
onnxruntime
opencv-python
numpy==1.26.4
pyyaml
